builds/codex-v101x-aeturnum-prime/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0.2
cryptography>=43.0.0
pyjwt[crypto]>=2.9.0
requests>=2.32.3
rich>=13.7.1
pydantic>=2.8.2
